10000 movement improvements, walk now dont make you fall on edges, using collisions of game actor

+++ 2DGame/02-Bubble/02-Bubble/GameActor.h
@@ -8,8 +8,8 @@ public:
 	int health;
 	int max_health;
 	bool down_key;
-	glm::vec2 size;
-	glm::vec2 position_col;
+	glm::ivec2 size;
+	glm::ivec2 position_col;
 	float magic = 16.0f;
 	float speed = 2.0f;
 protected:
